backend/requirements.txt
# FastAPI stack
fastapi==0.110.0
uvicorn==0.29.0
starlette==0.36.3
pydantic==2.6.4
pydantic-settings==2.2.1

# HTTP & utils
httpx==0.27.0
requests==2.31.0
python-dotenv==1.0.1

# Google Gemini AI
google-generativeai==0.4.1

# TensorFlow (compatible)
tensorflow==2.15.0

# Core dependencies
numpy==1.26.4
protobuf==4.25.3
typing-extensions>=4.8.0

# Optional but useful
python-multipart==0.0.9
rich==13.7.1
tqdm==4.66.2

pillow==10.3.0